+++ src/ts/main.ts
@@ -81,7 +81,7 @@ export {
 export { RecommendationOverlay } from './components/overlays/RecommendationOverlay';
 export { RecommendationItem, RecommendationItemConfig } from './components/RecommendationItem';
 export { SeekBarLabel, SeekBarLabelConfig } from './components/seekbar/SeekBarLabel';
-export { SubtitleOverlay } from './components/overlays/SubtitleOverlay';
+export { SubtitleOverlay, SubtitleOverlayConfig } from './components/overlays/SubtitleOverlay';
 export { SubtitleSelectBox } from './components/settings/SubtitleSelectBox';
 export { TitleBar, TitleBarConfig } from './components/TitleBar';
 export { VolumeControlButton, VolumeControlButtonConfig } from './components/buttons/VolumeControlButton';
@@ -117,8 +117,8 @@ export { WindowColorSelectBox } from './components/settings/subtitlesettings/Win
 export { WindowOpacitySelectBox } from './components/settings/subtitlesettings/WindowOpacitySelectBox';
 export { SubtitleSettingsResetButton } from './components/settings/subtitlesettings/SubtitleSettingsResetButton';
 export { ListBox, ListBoxConfig } from './components/lists/ListBox';
-export { SubtitleListBox } from './components/lists/SubtitleListBox';
-export { AudioTrackListBox } from './components/lists/AudioTrackListBox';
+export { SubtitleListBox, SubtitleListBoxConfig } from './components/lists/SubtitleListBox';
+export { AudioTrackListBox, AudioTrackListBoxConfig } from './components/lists/AudioTrackListBox';
 export { SettingsPanelPage, SettingsPanelPageConfig } from './components/settings/SettingsPanelPage';
 export { SettingsPanelPageBackButton } from './components/settings/SettingsPanelPageBackButton';
 export { SettingsPanelPageOpenButton } from './components/settings/SettingsPanelPageOpenButton';
@@ -137,6 +137,7 @@ export {
   ListSelector,
   ListSelectorConfig,
   ListItem,
+  ListItemComparator,
   ListItemFilter,
   ListItemLabelTranslator,
 } from './components/lists/ListSelector';